RT: small refactoring on fill_disruption
split fill_diruption_from_chaos to separate the navitia::Disruption
object creation and the change of the model
