+++ source/kraken/tests/CMakeLists.txt
@@ -9,11 +9,11 @@ target_link_libraries(disruption_reader_test workers data types pb_lib utils log
 ADD_BOOST_TEST(disruption_reader_test)
 
 add_executable(fill_disruption_from_chaos_tests fill_disruption_from_chaos_tests.cpp)
-target_link_libraries(fill_disruption_from_chaos_tests fill_disruption_from_chaos ed workers data types pb_lib utils log4cplus tcmalloc ${Boost_LIBRARIES} ${Boost_DATE_TIME_LIBRARY} protobuf)
+target_link_libraries(fill_disruption_from_chaos_tests make_disruption_from_chaos ed workers data types pb_lib utils log4cplus tcmalloc ${Boost_LIBRARIES} ${Boost_DATE_TIME_LIBRARY} protobuf)
 ADD_BOOST_TEST(fill_disruption_from_chaos_tests)
 
 add_executable(worker_test worker_test.cpp)
-target_link_libraries(worker_test ed workers data types pb_lib utils log4cplus tcmalloc ${Boost_LIBRARIES} ${Boost_DATE_TIME_LIBRARY} protobuf)
+target_link_libraries(worker_test make_disruption_from_chaos ed workers data types pb_lib utils log4cplus tcmalloc ${Boost_LIBRARIES} ${Boost_DATE_TIME_LIBRARY} protobuf)
 ADD_BOOST_TEST(worker_test)
 
 add_executable(realtime_test realtime_test.cpp)
